requirements.txt
langgraph
langchain-core
langchain-openai
langchain-experimental
langchain-text-splitters
langchain-huggingface
langchain-chroma

chromadb
unstructured
python-dotenv
tavily-python
ollama

streamlit
streamlit-nested-layout

pdfplumber
pyperclip
sentence-transformers